requirements.txt
livekit==0.18.1
livekit_agents==0.11.3
livekit_api==0.8.0
livekit_plugins_deepgram==0.6.11
livekit_plugins_elevenlabs==0.7.7
livekit_plugins_openai==0.10.7
livekit_plugins_silero==0.7.3
livekit_protocol==0.7.0
python-dotenv==1.0.1
